+++ TraceLd.PlotlySharp/TraceLd.PlotlySharp/NestedPayloadModels/PlotlyVector.cs
@@ -5,10 +5,10 @@ namespace TraceLd.PlotlySharp.NestedPayloadModels
     public class PlotlyVector
     {
         [JsonProperty("x")]
-        public int X { get; set; }
+        public double X { get; set; }
         [JsonProperty("y")]
-        public int Y { get; set; }
+        public double Y { get; set; }
         [JsonProperty("z")]
-        public int Z { get; set; }
+        public double Z { get; set; }
     }
 }
